move the multiop to a different module
