+++ src/treemap/mod.rs
@@ -2,6 +2,7 @@ use crate::RoaringBitmap;
 use std::collections::BTreeMap;
 
 mod fmt;
+mod multiops;
 mod util;
 
 // Order of these modules matters as it determines the `impl` blocks order in
